Part 1
Let's complete the first part of the assignment. First, create a class called HealthProfessional, which will serve as the base class. This class has a few important parts. There are instance variables like ID, name and department (). Then there is a default constructor, which is like a default initialisation setting. There is also a constructor that initialises all the instance variables with the parameters passed in. In addition, there is a method that prints out all these instance variables. Finally, although the assignment doesn't say specifically, we add Getter and Setter methods, which better follow the encapsulation principle of object-oriented programming.