Part 2
Today we will continue from the first day. We already have the HealthProfessional base class, and now we want to create a GeneralPractitioner class, which is a subclass of HealthProfessional. This subclass has its own special features. It has a new instance variable isGoodAtFamilyMedicine. It also has a default constructor and a constructor that initialises all the instance variables. Here's what you need to pay attention to: when initialising the subclass instance variables, you also need to initialise the base class instance variables by calling the base class constructor via super. Then there is a method that prints the details of the general practitioner, including type, ID, name, department, and whether or not they are good at family medicine. Finally, getter and setter methods are also added. And in the main method, we briefly test creating a general practitioner object and printing the information (although this part of the test code will be further improved in the future, let's first briefly see if the creation and printing functions are working properly).