Part 3
Let's create another subclass, Specialist, which also inherits from the HealthProfessional class. This class also has its own unique features. It has an instance variable, specialtyField. As before, there is a default constructor and a constructor that initialises all the instance variables. Here, too, the base class constructor is called via super to initialise the base class instance variables. Then there is a method that prints out the details of the specialist, including type, ID, name, department and area of expertise. Finally, the same getter and setter methods are added. And in the main method, update the test code to create a specialist object and print the information, and also create a general practitioner object and print the information, so that you can compare and see if the information for the different types of doctors is printed correctly.